+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 - GCDaySection.parse_json, GCDaySection.parse_csv
 - test suite
 
+## 0.0.9 - 2017-04-16
+
+### Added
+- ml linear regression analysis
+
 ## 0.0.8 - 2017-04-15
 
 ### Refactored